+++ backend/app/models/endpoint.model.go
@@ -9,7 +9,7 @@ type Endpoint struct {
 	Endpoint   string            `json:"endpoint" ch:"endpoint"`
 	Duration   time.Duration     `json:"duration" ch:"duration"`
 	RecordedAt time.Time         `json:"recordedAt" ch:"recorded_at"`
-	StatusCode int32             `json:"statusCode" ch:"status_code"`
+	StatusCode int16             `json:"statusCode" ch:"status_code"`
 	BodySize   int32             `json:"bodySize" ch:"body_size"`
 	ClientIP   string            `json:"clientIP" ch:"client_ip"`
 	Scope      map[string]string `json:"scope" ch:"scope"`
